requirements.txt
pyexifinfo==0.4.0
PyYAML==3.12